test-requirements.txt
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.25.2
aioresponses==0.7.6
faker==22.0.0
pytest-env==1.1.3
pytest-dotenv==0.5.2
discord.py==2.3.2
python-dotenv==1.0.0
line-bot-sdk==3.7.0